+++ moduliths-sample-spring-boot/src/main/java/com/acme/myproject/booking/data/BookingEntity.java
@@ -1,6 +1,6 @@
 package com.acme.myproject.booking.data;
 
-import com.acme.myproject.customer.Customer;
+import com.acme.myproject.customer.data.Customer;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
