+++ moduliths-sample-spring-boot/src/main/java/com/acme/myproject/booking/service/FlightDoesNotExistException.java
@@ -1,4 +1,4 @@
-package com.acme.myproject.booking.business;
+package com.acme.myproject.booking.service;
 
 class FlightDoesNotExistException extends RuntimeException {
