+++ moduliths-sample-spring-boot/src/test/java/com/acme/myproject/ModulesTest.java
@@ -1,0 +1,13 @@
+package com.acme.myproject;
+
+import de.olivergierke.moduliths.model.Modules;
+import org.junit.Test;
+
+public class ModulesTest {
+
+  @Test
+  public void verifyModules() {
+    new Modules().verify();
+  }
+
+}
