+++ moduliths-sample-spring-boot/src/test/java/com/acme/myproject/booking/service/BookingServiceTest.java
@@ -1,9 +1,9 @@
-package com.acme.myproject.booking.business;
+package com.acme.myproject.booking.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.acme.myproject.booking.data.BookingRepository;
-import com.acme.myproject.customer.CustomerRepository;
+import com.acme.myproject.customer.data.CustomerRepository;
 import com.acme.myproject.flight.service.FlightService;
 import com.acme.myproject.flight.web.FlightController;
 import de.olivergierke.moduliths.model.test.ModuleTest;
